stickerModels: split model and wrapper
Before this change, each model was a derivation of the model plus a
wrapper. This meant that in order to just use the model (e.g. in the
sticker Python module), sticker itself was also in the closure.

This change makes model an attribute set with the attributes `model`
and `wrapper`. `model` contains just the model, whereas `wrapper`
provides the command-line wrappers for the model.